add test

+++ main.go
@@ -7,6 +7,6 @@ import (
 
 func main() {
 	var v float64
-	err := lodash.Chain([]float32{2.34, 1.324}).Avg().Value(&v)
+	err := lodash.Chain([]float64{2.34, 1.324}).Avg().Value(&v)
 	fmt.Println(err, v)
 }
